pydata/requirements.txt
fastapi==0.110.3 
scikit-learn==1.2.2 
diffusers==0.27.2
invisible-watermark==0.2.0
transformers==4.40.1
accelerate==0.29.3
safetensors==0.4.3
boto3==1.34.96
uvicorn==0.29.0
python-dotenv==1.0.1
openai==1.25.0
pydantic===2.7.1
peft==0.10.0
